workers/requirements.txt
psycopg2
openai
python-dotenv
nltk
git+https://github.com/casics/nostril.git
Pillow
pytesseract
requests